ai_solution/requirements.txt
Flask
requests
python-dotenv
flask_cors
google-generativeai
flask-sqlalchemy
flask-login
flask-migrate
psycopg2-binary
werkzeug
psycopg2
flask_restx
jwt